refactor(expo-todomvc-sync-cf): migrate to new multi-store API
