+++ examples/expo-todomvc-sync-cf/src/components/Filters.tsx
@@ -1,13 +1,14 @@
-import { useQuery, useStore } from '@livestore/react'
+import { useQuery } from '@livestore/react'
 import type React from 'react'
 import { Pressable, StyleSheet, Text, View } from 'react-native'
 
 import { uiState$ } from '../livestore/queries.ts'
 import { events } from '../livestore/schema.ts'
+import { useAppStore } from '../livestore/store.ts'
 import type { Filter } from '../types.ts'
 
 export const Filters: React.FC = () => {
-  const { store } = useStore()
+  const store = useAppStore()
   const { filter } = useQuery(uiState$)
 
   const setFilter = (newFilter: Filter) => store.commit(events.uiStateSet({ filter: newFilter }))
