+++ examples/expo-todomvc-sync-cf/src/components/Todo.tsx
@@ -1,14 +1,14 @@
 import { MaterialIcons } from '@expo/vector-icons'
-import { useStore } from '@livestore/react'
 import type React from 'react'
 import { StyleSheet, Text, TouchableOpacity, View } from 'react-native'
 
 import type { tables } from '../livestore/schema.ts'
 import { events } from '../livestore/schema.ts'
+import { useAppStore } from '../livestore/store.ts'
 import { Checkbox } from './Checkbox.tsx'
 
 export const Todo: React.FC<typeof tables.todos.Type> = ({ id, text, completed }) => {
-  const { store } = useStore()
+  const store = useAppStore()
 
   const handleDeleteTodo = () => store.commit(events.todoDeleted({ id, deletedAt: new Date() }))
